requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
pymupdf==1.23.8
numpy==1.24.3
pandas==2.1.3
requests==2.31.0
tenacity==8.2.3
pyyaml==6.0.1
tiktoken==0.5.1
waitress==3.0.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
git+https://github.com/EndeeLabs/endee.git